chapter_003/src/main/java/ru/job4j/collection/PassportOffice.java
package ru.job4j.collection;

import java.util.HashMap;
import java.util.Map;

public class PassportOffice {
    private Map<String, Citizen> citizens = new HashMap<>();

    public boolean add(Citizen citizen) {
        boolean added = false;
        if (!citizens.containsKey(citizen.getPassport())) {
            citizens.put(citizen.getPassport(), citizen);
            added = true;
        }
        return added;
    }

    public Citizen findByPassport(String passport) {
        return citizens.get(passport);
    }
}
